Changed Point3D to doubles

+++ raytracer/utilities/Point3D.cpp
@@ -5,9 +5,9 @@
 
 Point3D::Point3D() : x{0}, y{0}, z{0} {}
 
-Point3D::Point3D(const float c) : x{c}, y{c}, z{c} {}
+Point3D::Point3D(const double c) : x{c}, y{c}, z{c} {}
 
-Point3D::Point3D(const float _x, const float _y, const float _z)
+Point3D::Point3D(const double _x, const double _y, const double _z)
     : x{_x}, y{_y}, z{_z} {}
 
 Point3D::Point3D(const std::array<double, 3>& ary)
@@ -25,17 +25,17 @@ Point3D Point3D::operator+(const Vector3D& v) const {
 
 Point3D Point3D::operator-(const Vector3D& v) const { return *this + -v; }
 
-Point3D Point3D::operator*(const float s) const {
+Point3D Point3D::operator*(const double s) const {
   return Point3D(x * s, y * s, z * s);
 }
 
-float Point3D::d_squared(const Point3D& p) const {
+double Point3D::d_squared(const Point3D& p) const {
   return (x - p.x) * (x - p.x) + (y - p.y) * (y - p.y) + (z - p.z) * (z - p.z);
 }
 
-float Point3D::distance(const Point3D& p) const { return sqrt(d_squared(p)); }
+double Point3D::distance(const Point3D& p) const { return sqrt(d_squared(p)); }
 
-Point3D operator*(const float a, const Point3D& pt) { return pt * a; }
+Point3D operator*(const double a, const Point3D& pt) { return pt * a; }
 
 Point3D Point3D::min(const Point3D& a, const Point3D& b) {
   return Point3D(std::min(a.x, b.x), std::min(a.y, b.y), std::min(a.z, b.z));
